Solution Day 06, part two
Actually we were lucky in part one. There was the bug that the guard just
did one step ahead right after turning right. For that step the collision
check was missing. We got quite far without it. Only the final test data
for part two lead to a slightly too small number of positions.