Solution Day 09, part one, finally
Input data that ends with free disk space was not handled correctly.
The test data in the description always ended with a file, but the actual
puzzle input does not.
Additionally, there was a problem with overlapping ranges of the data that
is still to fill up and the one that has to be moved.