Improve MutableRingList's performance
Use a custom implementation instead of a LinkedList for the MutableRingList.

In MarbleMania (day 9 of 2018) the list is rotated at most 7 steps at once
and items are added and removed within that range. So, while indices (pointing
to a standard list to which was delegated) might get very big, operations on
an actually rotated linked list are quite local.

This reduces the day 9's runtime from about half an hour to a few milliseconds.